tokenized cleanup seeder

+++ app/Http/Controllers/UserController.php
@@ -6,7 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use App\User;
-use Cookie;
+// use Cookie;
 
 class UserController extends Controller
 {
@@ -17,26 +17,12 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
-        $per_page = $request->input('per_page');
-        if (empty($per_page)) {
-            $cookie_value = $request->cookie('per_page');
-            $per_page = !empty($cookie_value) ? $cookie_value : 10;
-        }
-        Cookie::queue('per_page', $per_page);
-        return response()->json(User::paginate($per_page), 201);
-        // return User::paginate($per_page);
-        // return $cookie_value;
-        // return User::get();
-    }
-
-    public function users(Request $request)
-    {
-        $per_page = $request->input('per_page');
-        if (empty($per_page)) {
-            $cookie_value = $request->cookie('per_page');
-            $per_page = !empty($cookie_value) ? $cookie_value : 10;
-        }
-        Cookie::queue('per_page', $per_page);
+        $per_page = $request->input('per_page') ?: 10;
+        // if (empty($per_page)) {
+        //     $cookie_value = $request->cookie('per_page');
+        //     $per_page = !empty($cookie_value) ? $cookie_value : 10;
+        // }
+        // Cookie::queue('per_page', $per_page);
         return response()->json(User::paginate($per_page), 201);
         // return User::paginate($per_page);
         // return $cookie_value;
@@ -56,7 +42,24 @@ class UserController extends Controller
                 'message' => 'Your login credentials did not match any records.'
             ], 404);
         }
-        $token = $user->createToken('app-access')->plainTextToken;
+        $abilities = null;
+        switch ($user->role) {
+            case 'admin': {
+                $abilities = ['post:create', 'post:edit', 'post:delete'];
+                break;
+            }
+            case 'editor': {
+                $abilities = ['post:create', 'post:edit'];
+                break;
+            }
+            default: {
+                return response()->json([
+                    'error' => true,
+                    'message' => 'INVALID USER ROLE'
+                ], 201);
+            }
+        }
+        $token = $user->createToken('app-access', $abilities)->plainTextToken;
         return response()->json([
             'user' => $user,
             'token' => $token
@@ -77,6 +80,12 @@ class UserController extends Controller
             'message' => 'All tokens have been revoked.',
             'user' => $user
         ], 201);
+    }
+
+    public function signout() {
+        $user = $request->user();
+        // https://github.com/laravel/sanctum/issues/48
+        $user->currentAccessToken()->delete();
     }
 
     /**
@@ -105,8 +114,8 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $data = $request->validate($this->fields);
         $user = new User();
+        $data = $request->validate($this->fields);
         $user->name = $data['name'];
         $user->email = $data['email'];
         $user->password = Hash::make($data['password']);
@@ -177,7 +186,6 @@ class UserController extends Controller
         }
         $user->save();
         return response()->json($user, 201);
-        // return $user;
     }
 
     /**
@@ -189,8 +197,17 @@ class UserController extends Controller
     public function destroy($id)
     {
         $user = User::find($id);
-        $user->delete();
-        return response()->json($user, 201);
+        if ($user->role === 'admin') {
+            $user->delete();
+            return response()->json([
+                "user" => $user,
+                "message" => 'User successfully deleted'
+            ], 201);
+        }
+        return response()->json([
+            'error' => true,
+            'message' => 'You are not authorized to delete users'
+        ], 201);
         // return $user;
     }
 }
